refactor: split code into specialized manager classes
Improved code organization by separating FileManager and Settings
into focused manager classes. Updated dependencies.
