+++ package.json
@@ -19,19 +19,19 @@
 	"author": "LonoxX",
 	"license": "MIT",
 	"devDependencies": {
-		"@types/node": "^24.0.3",
-		"@typescript-eslint/eslint-plugin": "^8.34.1",
-		"@typescript-eslint/parser": "^8.34.1",
+		"@types/node": "^24.9.1",
+		"@typescript-eslint/eslint-plugin": "^8.46.2",
+		"@typescript-eslint/parser": "^8.46.2",
 		"builtin-modules": "^5.0.0",
-		"esbuild": "^0.25.5",
-		"eslint": "^9.29.0",
+		"esbuild": "^0.25.11",
+		"eslint": "^9.38.0",
 		"obsidian": "latest",
 		"tslib": "^2.8.1",
-		"typescript": "^5.8.3"
+		"typescript": "^5.9.3"
 	},
 	"dependencies": {
 		"date-fns": "^4.1.0",
-		"octokit": "^5.0.3",
-		"prettier": "^3.5.3"
+		"octokit": "^5.0.4",
+		"prettier": "^3.6.2"
 	}
 }
