+++ src/util/templateUtils.ts
@@ -379,70 +379,6 @@ export function createPullRequestTemplateData(
 }
 
 /**
- * Get available template variables as a help text
- * @returns String describing available template variables
- */
-export function getTemplateHelp(): string {
-	return `Available template variables:
-
-Basic Information:
-• {title} - Issue/PR title
-• {title_yaml} - Issue/PR title (YAML-escaped for use in frontmatter)
-• {number} - Issue/PR number
-• {status} / {state} - Current status (open, closed, etc.)
-• {author} - Username who created the issue/PR
-• {body} - Issue/PR description/body
-• {url} - Web URL
-• {repository} - Full repository name (owner/repo)
-• {owner} - Repository owner
-• {repoName} - Repository name only
-• {type} - "issue" or "pr"
-
-Assignees:
-• {assignee} - Primary assignee (first one if multiple)
-• {assignees} - All assignees as comma-separated list
-• {assignees_list} - All assignees as bulleted list
-• {assignees_yaml} - All assignees as YAML inline array [\"user1\", \"user2\"]
-
-Labels:
-• {labels} - All labels as comma-separated list
-• {labels_list} - All labels as bulleted list
-• {labels_hash} - All labels as hashtags (#label1 #label2)
-• {labels_yaml} - All labels as YAML inline array [\"enhancement\", \"bug\"]
-
-Dates:
-• {created} - Creation date
-• {updated} - Last update date
-• {closed} - Closed date (if closed)
-
-Pull Request Specific:
-• {mergedAt} - Merge date (if merged)
-• {mergeable} - Whether PR can be merged
-• {merged} - Whether PR is merged
-• {baseBranch} - Target branch
-• {headBranch} - Source branch
-
-Additional Info:
-• {milestone} - Milestone title
-• {commentsCount} - Number of comments
-• {isLocked} - Whether issue/PR is locked
-• {lockReason} - Lock reason (if locked)
-• {comments} - Formatted comments section (available only in content templates)
-
-Conditional Blocks:
-• {variable:content} - Shows content only if variable has a value
-• Example: {closed:- **Closed:** {closed}} - Shows "- **Closed:** [date]"
-• Example: {milestone:📌 Milestone: {milestone}} - Shows milestone info
-
-Examples:
-• "{title} - Issue {number}" → "Bug fix - Issue 123"
-• "{type} {number} - {title}" → "issue 123 - Bug fix"
-• "[{status}] {title} ({assignee})" → "[open] Bug fix (username)"
-• "{repoName}-{number} {title}" → "myproject-123 Bug fix"
-• "{closed:Closed on {closed}}" → "Closed on 2024-01-15" (only if closed)`;
-}
-
-/**
  * Extract the number from a filename based on the template that was used to create it
  * @param filename The filename to extract the number from
  * @param template The template that was used to create the filename
