+++ styles.css
@@ -1148,9 +1148,6 @@ button:disabled {
 	gap: 5px;
 	color: var(--text-accent);
 	font-weight: 500;
-	padding: 4px 10px;
-	border-radius: 4px;
-	background-color: rgba(var(--interactive-accent-rgb), 0.1);
 }
 
 .github-issues-empty-state {
@@ -1635,4 +1632,98 @@ button:disabled {
 	font-size: 0.85em;
 	text-align: center;
 	max-width: 500px;
+}
+
+/* Bulk Actions Toolbar */
+.github-issues-bulk-actions-toolbar {
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	padding: 12px 16px;
+	margin-bottom: 16px;
+	background-color: var(--background-secondary);
+	border: 1px solid var(--background-modifier-border);
+	border-radius: 6px;
+	gap: 16px;
+}
+
+.github-issues-bulk-action-info {
+	display: flex;
+	align-items: center;
+	gap: 8px;
+}
+
+.github-issues-selected-count {
+	font-weight: 600;
+	color: var(--text-normal);
+	font-size: 0.95em;
+}
+
+.github-issues-bulk-action-buttons {
+	display: flex;
+	gap: 8px;
+	align-items: center;
+}
+
+.github-issues-select-all-button,
+.github-issues-deselect-all-button {
+	padding: 6px 12px;
+	font-size: 0.9em;
+	background-color: var(--interactive-normal);
+	color: var(--text-normal);
+	border: 1px solid var(--background-modifier-border);
+	border-radius: 4px;
+	cursor: pointer;
+	transition: all 0.2s ease;
+}
+
+.github-issues-select-all-button:hover,
+.github-issues-deselect-all-button:hover {
+	background-color: var(--interactive-hover);
+}
+
+.github-issues-remove-selected-button {
+	display: inline-flex;
+	align-items: center;
+	gap: 6px;
+	padding: 6px 14px;
+	font-size: 0.9em;
+	cursor: pointer;
+	transition: all 0.2s ease;
+}
+
+.github-issues-remove-selected-button:disabled {
+	opacity: 0.5;
+	cursor: not-allowed;
+}
+
+/* Repository Checkbox */
+.github-issues-repo-checkbox {
+	cursor: pointer;
+	width: 18px;
+	height: 18px;
+	margin-right: 8px;
+	flex-shrink: 0;
+}
+
+/* Delete Modal Repository List */
+.github-issues-delete-repo-list {
+	margin: 12px 0;
+	padding: 12px;
+	background-color: var(--background-secondary);
+	border-radius: 4px;
+	max-height: 300px;
+	overflow-y: auto;
+}
+
+.github-issues-delete-repo-list ul {
+	margin: 0;
+	padding-left: 20px;
+	list-style-type: disc;
+}
+
+.github-issues-delete-repo-list li {
+	margin: 4px 0;
+	color: var(--text-normal);
+	font-size: 0.9em;
 }
